# integrate test suite