--- a/pom.xml
+++ b/pom.xml
@@ -149,6 +149,13 @@
 			<scope>test</scope>
 		</dependency>
 
+                <dependency>
+                    <groupId>com.squareup.okhttp3</groupId>
+                    <artifactId>okhttp</artifactId>
+                    <version>3.4.2</version>
+                    <scope>test</scope>
+                </dependency>
+
 	</dependencies>
 
 	<repositories>
@@ -216,8 +223,8 @@
 		</resources>
 	</build>
 	<properties>
-		<maven.compiler.source>1.6</maven.compiler.source>
-		<maven.compiler.target>1.6</maven.compiler.target>
+		<maven.compiler.source>1.8</maven.compiler.source>
+		<maven.compiler.target>1.8</maven.compiler.target>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 	</properties>