--- a/src/test/resources/integration-test-data/world/france/iledefrance/test_base.csv
+++ b/src/test/resources/integration-test-data/world/france/iledefrance/test_base.csv
@@ -1,0 +1,13 @@
+comment,query,expected_name,expected_housenumber,expected_street,expected_city,expected_postcode,expected_country,expected_osm_id,expected_osm_key,expected_osm_value,lon,lat,limit
+,courbevoie france,,,,,92400,France,,,administrative,,,
+,ile de france,,,,,,France,,,,,,
+,hauts de seine france,,,,,,France,,,,,,
+,rue des pyrénées paris france,Rue des Pyrénées,,,Paris,,France,,,,,,
+,eiffel tower paris,,,,Paris,,France,,,,,,
+,grands boulevards paris,,,,Paris,,France,,,,,,
+,cimetière du père lachaise,,,,Paris,75020,France,,,,,,
+,rue de lourmel paris,Rue de Lourmel,,,Paris,75015,France,,,,,,
+,25 rue bergère paris,,25,Rue Bergère,Paris,75009,France,,,,,,
+,rue bergère paris,Rue Bergère,,,Paris,75009,France,,,,,,
+Any rue de l'église,rue de l'église,Rue de l'Église,,,,,,,,,,,
+With center on Magny-le-Hongre,rue de l'église,Rue de l'Église,,,Magny-le-Hongre,,,,highway,,2.8131288,48.8651097,