--- a/src/test/resources/integration-test-data/world/france/iledefrance/test_fuzzy.csv
+++ b/src/test/resources/integration-test-data/world/france/iledefrance/test_fuzzy.csv
@@ -1,0 +1,14 @@
+query;expected_name;expected_street;expected_city;expected_postcode;expected_country;limit
+pairs france;Paris;;;75015;France;2
+pairs frnace;Paris;;;75015;France;2
+rue des bregers paris france;;Rue des Bergers;Paris;75015;France;2
+rue des bregers prais frnace;;Rue des Bergers;Paris;75015;France;2
+rue des begers pais franec;;Rue des Bergers;Paris;75015;France;2
+montreil;Montreuil;;;93100;France;2
+chateau ladon;Château-Landon;;;77570;France;2
+chateua landon;Château-Landon;;;77570;France;2
+chateua landno;Château-Landon;;;77570;France;2
+canal sant martin;Canal Saint-Martin;;;;France;2
+gare montparnase;Gare Montparnasse;;Paris;;France;2
+gare monparnasse;Gare Montparnasse;;Paris;;France;2
+allées paul favrier;Allée Paul Favier;;Champs-sur-Marne;;;2