--- a/src/test/resources/integration-test-data/world/france/iledefrance/test_long_queries.yml
+++ b/src/test/resources/integration-test-data/world/france/iledefrance/test_long_queries.yml
@@ -1,0 +1,23 @@
+39 rue du Caire, 75002 PARIS:
+    expected:
+        housenumber: 39
+        street: Rue du Caire
+        city: Paris
+
+16 Bis Boulevard Montmartre Paris 75009 France:
+    expected:
+        housenumber: 16b
+        street: Boulevard Montmartre
+        city: Paris
+
+130 rue de Lourmel – 75015 Paris, France:
+    expected:
+        housenumber: 130
+        street: Rue de Lourmel
+        city: Paris
+
+Immeuble Maille Nord II 8, avenue Montaigne 93160 Noisy-le-Grand:
+    expected:
+        housenumber: 8
+        street: Avenue Montaigne
+        city: Noisy-le-Grand