--- a/src/test/resources/integration-test-data/world/france/iledefrance/test_pois.yml
+++ b/src/test/resources/integration-test-data/world/france/iledefrance/test_pois.yml
@@ -1,0 +1,56 @@
+lycée henri IV:
+    expected:
+        name: Lycée et Collège Henri IV
+
+piscine pailleron:
+    expected:
+        name: Piscine Pailleron
+
+sacre coeur:
+    expected:
+        osm_id: 23762981
+
+basilique sacre coeur:
+    expected:
+        osm_id: 23762981
+
+numa paris:
+    expected:
+        osm_id: 55046167
+
+halte nautique de la Villette:
+    expected:   
+        osm_id: 2010316694
+
+port arsenal:
+    expected:
+        name: Port de l'Arsenal
+        city: Paris
+
+tour eiffel:
+    expected:
+        name: Tour Eiffel
+        city: Paris
+    lang: fr
+
+tour des nombres:
+    expected:
+        name: Tour des Nombres
+        city: Paris
+
+bibliothèque françois mitterrand:
+    expected:
+        osm_id: 56132291
+
+bibliothèque nationale de france:
+    expected:
+        osm_id: 56132291
+    size: 2
+    comment: there are two bibliothèques nationales
+
+cour carrée:
+    expected:
+        name: Cour Carrée
+        city: Paris
+    lon: 2.3672929240873075
+    lat: 48.849559049999996