--- a/src/test/resources/integration-test-data/world/france/iledefrance/test_suggest.csv
+++ b/src/test/resources/integration-test-data/world/france/iledefrance/test_suggest.csv
@@ -1,0 +1,1 @@
+comment;tried_query;tried_location;tried_language;expected_maximum_position;expected_location;expected_location_tolerance;expected_name;expected_type;expected_housenumber;expected_street;expected_city;expected_postcode;expected_country