--- a/src/test/resources/integration-test-data/world/france/test_hamlet.yml
+++ b/src/test/resources/integration-test-data/world/france/test_hamlet.yml
@@ -1,0 +1,5 @@
+beauregard montbrun-bocage:
+    expected:
+        name: Beauregard
+        osm_key: place
+        osm_value: hamlet