--- a/src/test/resources/integration-test-data/world/france/test_villages.yml
+++ b/src/test/resources/integration-test-data/world/france/test_villages.yml
@@ -1,0 +1,12 @@
+montbrun-bocage:
+    expected:
+        name: Montbrun-Bocage
+        postcode: 31310
+
+devesset:
+    expected:
+        name: Devesset
+
+fornex:
+    expected:
+        name: Fornex