… src/test/resources/integration-test-data/world/france/iledefrance/fuzzy/test_addresses_fuzzy.csv
… src/test/resources/integration-test-data/world/france/iledefrance/test_addresses.csv
… src/test/resources/integration-test-data/world/france/iledefrance/test_public_culture_buildings.csv
… src/test/resources/integration-test-data/world/france/iledefrance/test_public_culture_buildings_pois.csv
… src/test/resources/integration-test-data/world/france/nordpasdecalais/test_addresses.csv
… src/test/resources/integration-test-data/world/france/nordpasdecalais/test_postal_relays_addresses.csv
… src/test/resources/integration-test-data/world/france/nordpasdecalais/test_public_buildings_addresses.csv
… src/test/resources/integration-test-data/world/france/test_alsace.csv
… src/test/resources/integration-test-data/world/france/test_aquitaine.csv
… src/test/resources/integration-test-data/world/france/test_auvergne.csv
… src/test/resources/integration-test-data/world/france/test_bassenormandie.csv
… src/test/resources/integration-test-data/world/france/test_bourgogne.csv
… src/test/resources/integration-test-data/world/france/test_bretagne.csv
… src/test/resources/integration-test-data/world/france/test_centre.csv
… src/test/resources/integration-test-data/world/france/test_champagneardenne.csv
… src/test/resources/integration-test-data/world/france/test_franchecomte.csv
… src/test/resources/integration-test-data/world/france/test_hautenormandie.csv
… src/test/resources/integration-test-data/world/france/test_languedocroussillon.csv
… src/test/resources/integration-test-data/world/france/test_limousin.csv
… src/test/resources/integration-test-data/world/france/test_lorraine.csv
… src/test/resources/integration-test-data/world/france/test_midipyrenees.csv
… src/test/resources/integration-test-data/world/france/test_paysdelaloire.csv
… src/test/resources/integration-test-data/world/france/test_picardie.csv
… src/test/resources/integration-test-data/world/france/test_poitoucharentes.csv
… src/test/resources/integration-test-data/world/france/test_provencealpescotedazur.csv
… src/test/resources/integration-test-data/world/france/test_rhonealpes.csv
… src/test/resources/integration-test-data/world/germany/test_cities.csv
… src/test/resources/integration-test-data/world/germany/test_missing_housenumber.csv
… src/test/resources/integration-test-data/world/germany/test_regions.csv
… src/test/resources/integration-test-data/world/germany/test_towns.csv