This PR fixes #23 still some minor todos - see below.

The test suite was taken from https://github.com/geocoders/geocoder-tester but using a simple java class to execute the tests to avoid setting up separate tooling. Most of the test cases that ran before already run and I was able to compare the current PR #254 with master:

| test         | total request tests | failures 254 (time it took) | failures master (took)
|-------------|-------|------------------------|----------------
|Poland     | 120  | 50    (7s)                | 54 (18s)
|USA         | 300  |  84    (20s)             | 81 (47s)
|Germany | 351  | 52  (20s)                | 50 (54s)
|France1   | 145  | 77 (11s)                 | 64 (30s)
|France2   | 1810| 932 (110s)             | 867 (300s)

So the quality of the new PR is really comparable, only in France or some things to improve. The performance comparison is a bit unfair as I've done photon vs. some other setup but some smaller tests with the same setup still indicate a fair speed up.

TODOs:

 - [ ] add license to attribute work from @yohanboniface or I can also change the git author to him (if this would be ok)
 - [ ] read yml and json files
 - [ ] convert py code into json file
 - [ ] add issues we had while #254