@yohanboniface 2017-10-31
> The test suite was taken from https://github.com/geocoders/geocoder-tester but using a simple java class to execute the tests to avoid setting up separate tooling. 

That's a bit sad, the geocoders tools are made so we can factorize the efforts in maintaining tools around geocoders them-selves.
I can fix any issue in geocoder-tester to make it easier to use, if needed.