@karussell 2017-10-31
The problem is the sad state of python (why is python2 still around and not compatible with python3!?): I couldn't get python3 working as I need python2 for some other stuff here and a coexistence was above my linux foo. I can comment in the other repo if this helps.

Also I do think that we could share the data files and use different code to check this. For me as a Java dev I like having the different tests in my IDE and see kind of a summary and also that we could execute this on travis or somewhere else without much hassle.