@yohanboniface 2017-10-31
> I couldn't get python3 working as I need python2 for some other stuff here and a coexistence was above my linux foo. 

It's a matter of using virtualenvs. :)

> I can comment in the other repo if this helps.

Yes, please. I can provide a complete recipe for non python devs.