@karussell 2017-10-31
> It's a matter of using virtualenvs. :)

That was what I tried but couldn't get it working faster then writing some small Java code (which I also could slightly adapt to benchmark the GH API that currently has a custom JSON format).