@yohanboniface 2017-10-31
See http://docs.python-guide.org/en/latest/dev/virtualenvs/

I can also setup a Docker image if that helps.

> writing some small Java code

The first bits are always small and easy to setup. ;) As you known, the FLOSS hard bit is maintaining in the long term. :)