@karussell 2017-10-31
> The first bits are always small and easy to setup. ;) As you known, the FLOSS hard bit is maintaining in the long term. :)

I know. But it is obvious that for others the test suite wasn't as accessible as it should be and so too much decoupling is also a risc :). So we should improve this, and as the test data is really the important bit here maybe we can download this on the fly and have it "directly integrated".