@karussell 2017-11-06
Another problem is if we wanted to run this regularly. Then we need to find a way to ignore certain failures but not all. Problem is mentioned in more detail here https://github.com/geocoders/geocoder-tester/issues/20