@karussell 2018-01-20
Will try to avoid duplicating this here and enhancing the geocoder-tester